Database + UI Big update KB

+++ app/src/main/java/com/example/edtcarboncounter/data/transportObject.kt
@@ -3,4 +3,4 @@ package com.example.edtcarboncounter.data
 //deleted = 1
 //not deleted = 0
 
-class transportObject (var type: String, var Sdistance: String, var Ldistance: Long, var deleted: Int)
+class transportObject (var type: String, var Sdistance: String, var Ldistance: Double, var deleted: Int)
